Fix bug in root unit computations
Replace computed rootUnitItems with useRootUnits composable
