+++ src/components/ImportMilxStep.vue
@@ -7,7 +7,6 @@ import { injectStrict, nanoid, removeUndefined } from "@/utils";
 import { activeScenarioKey } from "@/components/injects";
 import type { NUnit } from "@/types/internalModels";
 import type { Point } from "geojson";
-import type { SymbolItem } from "@/types/constants";
 import SymbolCodeSelect from "@/components/SymbolCodeSelect.vue";
 import { setCharAt } from "@/components/helpers";
 import { SID_INDEX } from "@/symbology/sidc";
@@ -20,6 +19,7 @@ import { propReduce } from "@turf/meta";
 import { featureCollection } from "@turf/helpers";
 import { pick } from "es-toolkit";
 import { Button } from "@/components/ui/button";
+import { useRootUnits } from "@/composables/scenarioUtils.ts";
 
 interface Props {
   data: MilxImportedLayer[];
@@ -29,18 +29,10 @@ const props = defineProps<Props>();
 const emit = defineEmits(["cancel", "loaded"]);
 const { unitActions, store: scnStore, time } = injectStrict(activeScenarioKey);
 const store = useImportStore();
-const { state } = scnStore;
 
 const selectedUnits = ref<FlatItem[]>([]);
 
-const rootUnitItems = computed((): SymbolItem[] => {
-  return Object.values(state.sideGroupMap)
-    .map((value) => value.subUnits)
-    .flat()
-    .map((e) => state.unitMap[e])
-    .map((u) => ({ text: u.name, code: u.id, sidc: u.sidc }));
-});
-
+const { rootUnitItems } = useRootUnits();
 const parentUnitId = ref(rootUnitItems.value[0].code as string);
 
 interface FlatItem extends ImportGeoJsonFeature {
